Rename with_request to with_reqwest modules
