+++ Cargo.toml
@@ -22,7 +22,7 @@ reqwest = { version = "0.11", default-features = false, optional = true }
 log = { version = "0.4", default-features = false, optional = true }
 
 [features]
-default = ["non_blocking"]
+default = ["with_reqwest", "prometheus_client_crate"]
 non_blocking = ["async-trait"]
 blocking = []
 prometheus_crate = ["prometheus"]
