+++ src/prometheus_crate.rs
@@ -19,7 +19,7 @@ use crate::non_blocking::MetricsPusher;
 #[cfg(feature = "with_reqwest")]
 use crate::non_blocking::Push;
 #[cfg(feature = "with_reqwest")]
-use crate::with_request::PushClient;
+use crate::with_reqwest::PushClient;
 #[cfg(feature = "with_reqwest")]
 use reqwest::Client;
 
@@ -117,7 +117,7 @@ where
 
 #[cfg(feature = "with_reqwest_blocking")]
 pub type PrometheusMetricsPusherBlocking = blocking::MetricsPusher<
-    blocking::with_request::PushClient,
+    blocking::with_reqwest::PushClient,
     PrometheusMetricsConverter,
     Vec<MetricFamily>,
     Vec<Box<dyn Collector>>,
@@ -132,12 +132,12 @@ where
 {
     /// Creates a new [`MetricsPusher`] with the given [`reqwest::blocking::Client`] client
     /// and the Url of your pushgateway instance.
-    pub fn blocking_from(
+    pub fn from(
         client: reqwest::blocking::Client,
         url: &Url,
     ) -> Result<PrometheusMetricsPusherBlocking> {
         blocking::MetricsPusher::new(
-            blocking::with_request::PushClient::new(client),
+            blocking::with_reqwest::PushClient::new(client),
             PrometheusMetricsConverter,
             url,
         )
